Refine bucketlist details UI. Implement promises at delete items and create functions in BucketlistDetailController.

+++ api/static/js/controllers/BucketlistDetailsController.js
@@ -13,7 +13,24 @@ BucketlistApp.controller('BucketlistDetailsController',
 					name: $scope.new_item_name,
 					done: $scope.new_item_status
 				};
-				BucketlistFactory.Item.create(data);
+				BucketlistFactory.Item.create(data)
+				.$promise.then(
+					function (response) {
+						$scope.bucketlist = BucketlistFactory.Bucketlist.getOne({buck_id: $routeParams.buckId});
+						var $toastContent = $('<strong style="color: #4db6ac;">Bucketlist created successfully.</strong>');
+						Materialize.toast($toastContent, 5000);
+						$scope.new_item_name = "";
+						$scope.new_item_status = "";
+					},
+					function (error) {
+						var $toastContent = $('<strong style="color: #f44336;">Error Creating bucketlist.</strong>');
+						Materialize.toast($toastContent, 5000);
+					}
+				);
+			}
+			else {
+				var $toastContent = $('<strong style="color: #f44336;">Item name not provided.</strong>');
+				Materialize.toast($toastContent, 5000);
 			}
 
 		};
@@ -23,11 +40,26 @@ BucketlistApp.controller('BucketlistDetailsController',
 		}
 
 		$scope.deleteItem = function (item_id) {
-			var data = {
-				buck_id: $routeParams.buckId,
-				item_id: item_id
-			};
-			BucketlistFactory.ItemDetail.deleteItem(data);
+			var conf = confirm('Proceed and delete?');
+
+			if (conf) {
+				var data = {
+					buck_id: $routeParams.buckId,
+					item_id: item_id
+				};
+				BucketlistFactory.ItemDetail.deleteItem(data)
+				.$promise.then(
+					function (response) {
+						var $toastContent = $('<strong style="color: #4db6ac;">Bucketlist successfully deleted.</strong>');
+						Materialize.toast($toastContent, 5000);
+						$scope.bucketlist = BucketlistFactory.Bucketlist.getOne({buck_id: $routeParams.buckId});
+					}, function (error) {
+						var $toastContent = $('<strong style="color: #f44336;">Bucketlist not edited.</strong>');
+						Materialize.toast($toastContent, 5000);
+					}
+				);
+			}
+
 		}
 
 	}
